+++ MatchesRest/src/main/java/es/iesjandula/statsbomb/matches_rest/stats/utils/ISeasonRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @version 1.0.0
  * @since 02/02/2023
  */
-public interface ISeasonRepository extends JpaRepository<Season, Long>
+public interface ISeasonRepository extends JpaRepository<Season, Integer>
 {
 }
